main.c
Demonstrates use of threads in C to create the illusion of passengers waiting in line at a roller coaster ride. There are 6 passengers in the example and only one car with 2 seats, which creates a race condition. Each thread fights for the car and the car locks when it has two passengers. After sleeping for two seconds the car is finished and allows the passengers to exit and then new ones to enter. This is done until all passengers have gone once.